+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## Unreleased
 
-##  v0.2.12
+##  v0.2.13
 
 - Remove deprecate attendeeIdentifiers from courseGroups source attribute
 - Add availablePlaces to courseGroups source attribute
 - Add currentSemesterKey and nextSemesterKey local data attributes to BaseCourse entity
+
+## v0.2.12
+
+- use campusonline API token cache
 
 ## v0.2.11